Add files via upload
This is my latest animation I've made so far. I call it  episode one 😂. I am currently working on episode two which is not related to episode one by any means possible🥲. I would have been done with episode two but school is putting a lot of load on me.